fixing errors

+++ Board.h
@@ -17,7 +17,7 @@ class Board
         long int WP,WN,WB,WR,WQ,WK,BP,BN,BB,BR,BQ,BK;
         char chessBoard[8][8];
         void arraytoBitBoard();
-        int convertStringToBitBoard(std::string bin_num);
+        long convertStringToBitBoard(std::string bin_num);
 
 
     public:
